requirements.txt
blinker==1.8.2
Bootstrap-Flask==2.2.0
click==8.1.7
colorama==0.4.6
Flask==2.3.2
future==1.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mime==0.1.0
python-dotenv==1.0.1
Werkzeug==3.0.3
WTForms==3.1.2
gunicorn==21.2.0
psycopg2-binary==2.9.9